Refactor: repository impl 관련 로직 infrastructure에서 분리
